Transmission and checks of firing
Method         internal void UpdatePlayerCrosshair(Client sender, dynamic crosshair)
needs work

+++ Tie Server/Game.cs
@@ -13,9 +13,9 @@ namespace Tie_Server
     /// <summary>
     /// Class that contains a game. It manages the gathering of players, the 
     /// </summary>
-    
-    public enum GameStatus { Lobby, Running, Finished}
-    public class Game: IDataReceiver
+
+    public enum GameStatus { Lobby, Running, Finished }
+    public class Game : IDataReceiver
     {
         public GameManager gameManager;
         public GameStatus gameStatus { get; set; } = GameStatus.Lobby;
@@ -34,7 +34,7 @@ namespace Tie_Server
             data.data = "Player " + newPlayer.name + " has joined the lobby";
             foreach (Player player in gameManager.players)
                 player.client.Write(data);
-            this.Start();
+            //this.Start();
         }
 
         public void handlePacket(dynamic data, Client sender)
@@ -42,7 +42,7 @@ namespace Tie_Server
             switch ((string)data.type)
             {
                 case "crosshair":
-                    gameManager.UpdatePlayerCrosshair(data.data.clientID, data.data.crosshair);
+                    gameManager.UpdatePlayerCrosshair(sender, data.data);
                     break;
                 case "highscorerequest":
                     Program.handleHighscoreRequest(sender);
@@ -68,7 +68,7 @@ namespace Tie_Server
                 client.Write(gameManager.GetGameData());
             }
         }
-            public void Start()
+        public void Start()
         {
             gameStatus = GameStatus.Running;
             var timerDelegate = new System.Timers.Timer(30000);
@@ -87,7 +87,8 @@ namespace Tie_Server
             List<HighScore> scores = GetHighScoresFromFile();
             scores.Add(GetHighestScore());
             scores.Sort();
-            scores.RemoveRange(10, scores.Count - 10); // trim so only 10 remain
+            if (scores.Count > 10)
+                scores.RemoveRange(10, scores.Count - 10); // trim so only 10 remain
             writeHighscoresToFile(scores);
         }
 
@@ -134,6 +135,6 @@ namespace Tie_Server
             }
         }
 
-       
+
     }
 }
